requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==4.2
django-filter==25.1
djangorestframework==3.16.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.8
msgpack==1.1.0
outcome==1.3.0.post0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
PySocks==1.7.1
redis==5.2.1
selenium==4.31.0
service-identity==24.2.0
setuptools==78.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
